@@ -2,32 +2,36 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router';
 
-import { trackDiscovery } from '../../utils/utils.js';
-import appConfig from '../../../../appConfig.js';
+import { trackDiscovery } from '../../utils/utils';
+import appConfig from '../../../../appConfig';
 
-const baseUrl = appConfig.baseUrl;
+const {
+  baseUrl,
+  displayTitle,
+} = appConfig;
 
 const Breadcrumbs = ({ query = '', type, bibUrl, itemUrl, edd }) => {
-  const defaultText = appConfig.displayTitle;
-  const onClick = (pageTitle) => trackDiscovery('Breadcrumbs', pageTitle);
-  const homeLink = (<li key="home">
-    <Link to={`${baseUrl}/`} onClick={() => onClick(defaultText)}>{defaultText}</Link>
-  </li>);
+  const defaultText = displayTitle;
+  const onClick = pageTitle => trackDiscovery('Breadcrumbs', pageTitle);
+  const homeLink = (
+    <li key="home">
+      <Link to={`${baseUrl}/`} onClick={() => onClick(defaultText)}>
+        {defaultText}
+      </Link>
+    </li>);
 
   /*
    * getCrumbs()
    * Returns an array of list elements to render for the breadcrumbs navigation.
    * This will keep pushing new elements onto an array until it reaches the spot desired.
-   * It will then return the list up to that point with the desired spot being a simple
+   * It will then return the list up to that point with the active location being a simple
    * text list element.
    * @returns {array}
    */
   const getCrumbs = () => {
     // The first link is the homepage and it will being appearing starting from the
     // Search Results page.
-    const crumbs = [
-      homeLink,
-    ];
+    const crumbs = [homeLink];
 
     if (type === 'search') {
       crumbs.push(<li key="search">Search Results</li>);
@@ -39,26 +43,27 @@ const Breadcrumbs = ({ query = '', type, bibUrl, itemUrl, edd }) => {
         <Link to={`${baseUrl}/search?${query}`} onClick={() => onClick('Search Results')}>
           Search Results
         </Link>
-      </li>
-    );
+      </li>);
 
     if (type === 'bib') {
       crumbs.push(<li key="bib">Item Details</li>);
       return crumbs;
     }
 
-    crumbs.push(<li key="bib">
-      <Link to={`${baseUrl}${bibUrl}`} onClick={() => onClick('Item Details')}>Item Details</Link>
-    </li>);
+    crumbs.push(
+      <li key="bib">
+        <Link to={`${baseUrl}${bibUrl}`} onClick={() => onClick('Item Details')}>Item Details</Link>
+      </li>);
 
     if (type === 'hold') {
       crumbs.push(<li key="hold">Item Request</li>);
       return crumbs;
     }
 
-    crumbs.push(<li key="hold">
-      <Link to={`${baseUrl}${itemUrl}`} onClick={() => onClick('Item Request')}>Item Request</Link>
-    </li>);
+    crumbs.push(
+      <li key="hold">
+        <Link to={`${baseUrl}${itemUrl}`} onClick={() => onClick('Item Request')}>Item Request</Link>
+      </li>);
 
     if (type === 'edd') {
       crumbs.push(<li key="edd">Electronic Delivery Request</li>);
@@ -75,8 +80,7 @@ const Breadcrumbs = ({ query = '', type, bibUrl, itemUrl, edd }) => {
           >
             Electronic Delivery Request
           </Link>
-        </li>
-      );
+        </li>);
     }
 
     // The last possible point in the breadcrumbs will be the Confirmation page.
