@@ -1,21 +1,21 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { validate } from '../../utils/formValidationUtils';
-import appConfig from '../../../../appConfig';
-
 import {
   mapObject as _mapObject,
   extend as _extend,
   isEmpty as _isEmpty,
 } from 'underscore';
 
+import { validate } from '../../utils/formValidationUtils';
+import appConfig from '../../../../appConfig';
+
 class ElectronicDeliveryForm extends React.Component {
   constructor(props) {
     super(props);
 
-    // this.props.form and this.props.error are coming from the server only in the
-    // no-js scenario. If they're not available, then we use this 'fallback', but the
-    // empty object structure is needed.
+    // this.props.form and this.props.error are coming from the server only
+    // in the no-js scenario. If they're not available, then we use this
+    // 'fallback', but the empty object structure is needed.
     this.state = {
       form: !_isEmpty(this.props.form) ? this.props.form :
         {
@@ -51,9 +51,10 @@ class ElectronicDeliveryForm extends React.Component {
   }
 
   handleUpdate(e, input) {
-    // Kind of hard to read. Basically, the `form` property is being updated and all
-    // the values are being retained. If we don't `extend` the object value for `form`,
-    // then only the last value in the form gets updated and the rest are gone.
+    // Kind of hard to read. Basically, the `form` property is being updated
+    // and all the values are being retained. If we don't `extend` the object
+    // value for `form`, then only the last value in the form gets updated
+    // and the rest are gone.
     this.setState({ form: _extend(this.state.form, { [input]: e.target.value }) });
   }
 
@@ -69,16 +70,16 @@ class ElectronicDeliveryForm extends React.Component {
       errorClass[key] = this.state.error[key] ? 'nypl-field-error' : '';
     });
 
-    // A lot of this can be refactored to be in a loop but that's a later and next step.
-    // I was thinking each `nypl-text-field` or `nypl-year-field` div can be
-    // its own component in a loop with the required props and errors passed down.
+    // A lot of this can be refactored to be in a loop but that's a later and
+    // next step. I was thinking each `nypl-text-field` or `nypl-year-field`
+    // div can be its own component in a loop with the required props and
+    // errors passed down.
     return (
       <form
         className="place-hold-form form electronic-delivery-form"
         action={`${appConfig.baseUrl}/edd`}
         method="POST"
-        onSubmit={(e) => this.submit(e)}
-        id="edd-request"
+        onSubmit={e => this.submit(e)}
       >
         <fieldset className="nypl-fieldset">
           <legend>
@@ -97,7 +98,7 @@ class ElectronicDeliveryForm extends React.Component {
                   aria-required="true"
                   name="emailAddress"
                   value={this.state.form.emailAddress}
-                  onChange={(e) => this.handleUpdate(e, 'emailAddress')}
+                  onChange={e => this.handleUpdate(e, 'emailAddress')}
                 />
                 <span
                   className="nypl-field-status"
@@ -111,7 +112,11 @@ class ElectronicDeliveryForm extends React.Component {
                   }
                 </span>
               </div>
-              <span>You may request a maximum of 50 pages.</span>
+              <span>You may request a maximum of 50 pages.</span><br />
+              <span>
+                Read <a href="https://www.nypl.org/help/request-research-materials#EDD">
+                more about this service</a>.
+              </span>
               <div className={`nypl-text-field ${errorClass.startPage}`}>
                 <label htmlFor="startPage" id="startPage-label">Starting Page Number
                   <span className="nypl-required-field">&nbsp;Required</span>
@@ -124,7 +129,7 @@ class ElectronicDeliveryForm extends React.Component {
                   aria-labelledby="startPage-label startPage-status"
                   name="startPage"
                   value={this.state.form.startPage}
-                  onChange={(e) => this.handleUpdate(e, 'startPage')}
+                  onChange={e => this.handleUpdate(e, 'startPage')}
                 />
                 <span
                   className="nypl-field-status"
@@ -137,7 +142,7 @@ class ElectronicDeliveryForm extends React.Component {
                       errorClass.startPage ?
                       this.state.error.startPage : 'Enter the first page of your selection.'
                     }
-                    </span>
+                  </span>
                 </span>
               </div>
 
@@ -153,7 +158,7 @@ class ElectronicDeliveryForm extends React.Component {
                   aria-labelledby="endPage-label endPage-status"
                   name="endPage"
                   value={this.state.form.endPage}
-                  onChange={(e) => this.handleUpdate(e, 'endPage')}
+                  onChange={e => this.handleUpdate(e, 'endPage')}
                 />
                 <span
                   className="nypl-field-status"
@@ -181,7 +186,7 @@ class ElectronicDeliveryForm extends React.Component {
                   aria-labelledby="chapterTitle-label chapterTitle-status"
                   name="chapterTitle"
                   value={this.state.form.chapterTitle}
-                  onChange={(e) => this.handleUpdate(e, 'chapterTitle')}
+                  onChange={e => this.handleUpdate(e, 'chapterTitle')}
                 />
                 <span
                   className="nypl-field-status"
@@ -216,7 +221,7 @@ class ElectronicDeliveryForm extends React.Component {
                   aria-labelledby="author-label"
                   name="author"
                   value={this.state.form.author}
-                  onChange={(e) => this.handleUpdate(e, 'author')}
+                  onChange={e => this.handleUpdate(e, 'author')}
                 />
               </div>
 
@@ -228,7 +233,7 @@ class ElectronicDeliveryForm extends React.Component {
                   aria-labelledby="date-label"
                   name="date"
                   value={this.state.form.date}
-                  onChange={(e) => this.handleUpdate(e, 'date')}
+                  onChange={e => this.handleUpdate(e, 'date')}
                 />
               </div>
 
@@ -240,7 +245,7 @@ class ElectronicDeliveryForm extends React.Component {
                   aria-labelledby="volume-label"
                   name="volume"
                   value={this.state.form.volume}
-                  onChange={(e) => this.handleUpdate(e, 'volume')}
+                  onChange={e => this.handleUpdate(e, 'volume')}
                 />
               </div>
 
@@ -252,11 +257,17 @@ class ElectronicDeliveryForm extends React.Component {
                   aria-labelledby="issue-label"
                   name="issue"
                   value={this.state.form.issue}
-                  onChange={(e) => this.handleUpdate(e, 'issue')}
+                  onChange={e => this.handleUpdate(e, 'issue')}
                 />
               </div>
-              <div className="nypl-text-field">
+              <div className="nypl-text-area-with-label">
                 <label htmlFor="requestNotes" id="requestNotes-label">Notes</label>
+                <p>
+                  Provide additional instructions here. In case electronic delivery is unavailable,
+                  please indicate a physical delivery location in the field below. For more
+                  information on placing an electronic delivery request, please
+                  see <a href="https://www.nypl.org/help/request-research-materials">Requesting Research Materials</a>.
+                </p>
                 <textarea
                   className="nypl-text-area"
                   id="requestNotes"
@@ -264,7 +275,7 @@ class ElectronicDeliveryForm extends React.Component {
                   aria-labelledby="requestNotes-label"
                   name="requestNotes"
                   value={this.state.form.requestNotes}
-                  onChange={(e) => this.handleUpdate(e, 'requestNotes')}
+                  onChange={e => this.handleUpdate(e, 'requestNotes')}
                 />
               </div>
             </div>
@@ -300,7 +311,6 @@ ElectronicDeliveryForm.propTypes = {
   bibId: PropTypes.string,
   itemId: PropTypes.string,
   itemSource: PropTypes.string,
-  pickupLocation: PropTypes.string,
   error: PropTypes.object,
   form: PropTypes.object,
   defaultEmail: PropTypes.string,
