@@ -1,7 +1,6 @@
+/* global window document */
 import React from 'react';
 import PropTypes from 'prop-types';
-import PatronStore from '../../stores/PatronStore.js';
-import appConfig from '../../../../appConfig.js';
 import { Link } from 'react-router';
 import {
   isArray as _isArray,
@@ -10,8 +9,10 @@ import {
 } from 'underscore';
 import DocumentTitle from 'react-document-title';
 
-import Breadcrumbs from '../Breadcrumbs/Breadcrumbs.jsx';
-import { trackDiscovery } from '../../utils/utils.js';
+import PatronStore from '../../stores/PatronStore';
+import appConfig from '../../../../appConfig';
+import Breadcrumbs from '../Breadcrumbs/Breadcrumbs';
+import { trackDiscovery } from '../../utils/utils';
 
 class HoldConfirmation extends React.Component {
   constructor(props) {
@@ -24,6 +25,8 @@ class HoldConfirmation extends React.Component {
 
   componentDidMount() {
     this.requireUser();
+
+    document.getElementById('confirmation-title').focus();
   }
 
   /**
@@ -38,19 +41,18 @@ class HoldConfirmation extends React.Component {
     }
 
     const fullUrl = encodeURIComponent(window.location.href);
-
     window.location.replace(`${appConfig.loginUrl}?redirect_uri=${fullUrl}`);
 
     return false;
   }
 
   /**
-   * goRestart(e)
+   * backToHome(e)
    * @param {event}
    * Renders the route back to home page for single page application implementation.
    *
    */
-  goRestart(e) {
+  backToHome(e) {
     e.preventDefault();
 
     trackDiscovery('Discovery Search', 'New Search');
@@ -58,12 +60,12 @@ class HoldConfirmation extends React.Component {
   }
 
   /**
-   * goSearchResults(e)
+   * goToSearchResults(e)
    * @param {event}
    * Renders the route back to search results page.
    *
    */
-  goSearchResults(e) {
+  goToSearchResults(e) {
     e.preventDefault();
 
     trackDiscovery('Discovery Search', 'Existing Search');
@@ -132,7 +134,7 @@ class HoldConfirmation extends React.Component {
     if (this.props.location.query.fromUrl) {
       return (
         <span id="go-to-shared-catalog"> You may also try your search in
-          our <Link to={`${appConfig.baseUrl}/`} onClick={(e) => this.goRestart(e)}>Shared
+          our <Link to={`${appConfig.baseUrl}/`} onClick={e => this.backToHome(e)}>Shared
           Collection Catalog</Link>.
         </span>
       );
@@ -149,7 +151,7 @@ class HoldConfirmation extends React.Component {
         <Link
           id="start-new-search"
           to={`${appConfig.baseUrl}/`}
-          onClick={(e) => this.goRestart(e)}
+          onClick={e => this.backToHome(e)}
         >
           {text}
         </Link>.
@@ -198,7 +200,7 @@ class HoldConfirmation extends React.Component {
         // We use this.props.location.query.searchKeywords here for the query from
         // the URL to deal with no js situation.
         to={`${appConfig.baseUrl}/search?q=${this.props.location.query.searchKeywords}`}
-        onClick={(e) => this.goSearchResults(e)}
+        onClick={e => this.goToSearchResults(e)}
       >
         Go back to your search results
       </Link> or </span>
@@ -231,7 +233,7 @@ class HoldConfirmation extends React.Component {
     if (this.props.deliveryLocations && this.props.deliveryLocations.length) {
       if (pickupLocation !== 'edd') {
         deliveryLocation = _findWhere(
-          this.props.deliveryLocations, { '@id': `loc:${pickupLocation}` }
+          this.props.deliveryLocations, { '@id': `loc:${pickupLocation}` },
         );
       } else {
         deliveryLocation = {
@@ -303,7 +305,7 @@ class HoldConfirmation extends React.Component {
                     itemUrl={`/hold/request/${bibId}-${itemId}`}
                     edd={pickupLocation === 'edd'}
                   />
-                  <h1>{confirmationPageTitle}</h1>
+                  <h1 id="confirmation-title" tabIndex="0">{confirmationPageTitle}</h1>
                 </div>
               </div>
             </div>
