@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { Link } from 'react-router';
 import { isEmpty as _isEmpty } from 'underscore';
 
-import appConfig from '../../../../appConfig.js';
+import appConfig from '../../../../appConfig';
 
 const ItemTableRow = ({ item, bibId, getRecord, searchKeywords }) => {
   if (_isEmpty(item)) {
@@ -25,7 +25,7 @@ const ItemTableRow = ({ item, bibId, getRecord, searchKeywords }) => {
           to={
             `${appConfig.baseUrl}/hold/request/${bibId}-${item.id}?searchKeywords=${searchKeywords}`
           }
-          onClick={(e) => getRecord(e, bibId, item.id)}
+          onClick={e => getRecord(e, bibId, item.id)}
           tabIndex="0"
         >
           Request
