@@ -13,7 +13,7 @@ import {
   trackDiscovery,
 } from '../../utils/utils';
 import ItemTable from '../Item/ItemTable';
-import appConfig from '../../../../appConfig.js';
+import appConfig from '../../../../appConfig';
 
 class ResultsList extends React.Component {
   constructor(props) {
@@ -40,24 +40,21 @@ class ResultsList extends React.Component {
     ajaxCall(`${appConfig.baseUrl}/api/bib?bibId=${bibId}`,
       (response) => {
         Actions.updateBib(response.data);
-        setTimeout(
-          () => { this.props.updateIsLoadingState(false); },
-          500
-        );
-
-        this.routeHandler(`${appConfig.baseUrl}/bib/${bibId}`);
+        setTimeout(() => {
+          this.props.updateIsLoadingState(false);
+          this.routeHandler(`${appConfig.baseUrl}/bib/${bibId}`);
+        }, 500);
       },
-      error => {
-        setTimeout(
-          () => { this.props.updateIsLoadingState(false); },
-          500
-        );
+      (error) => {
+        setTimeout(() => {
+          this.props.updateIsLoadingState(false);
+        }, 500);
 
         console.error(
           'Error attempting to make an ajax request to fetch a bib record from ResultsList',
-          error
+          error,
         );
-      }
+      },
     );
   }
 
@@ -80,23 +77,21 @@ class ResultsList extends React.Component {
         Actions.updateBib(response.data.bib);
         Actions.updateDeliveryLocations(response.data.deliveryLocations);
         Actions.updateIsEddRequestable(response.data.isEddRequestable);
-        setTimeout(
-          () => { this.props.updateIsLoadingState(false); },
-          500
-        );
-
-        this.routeHandler(`${appConfig.baseUrl}/hold/request/${bibId}-${itemId}`);
+        setTimeout(() => {
+          this.props.updateIsLoadingState(false);
+          this.routeHandler(`${appConfig.baseUrl}/hold/request/${bibId}-${itemId}`);
+        }, 500);
       },
-      error => {
-        setTimeout(
-          () => { this.props.updateIsLoadingState(false); },
-          500
-        );
+      (error) => {
+        setTimeout(() => {
+          this.props.updateIsLoadingState(false);
+        }, 500);
+
         console.error(
           'Error attemping to make an ajax request to fetch an item in ResultsList',
-          error
+          error,
         );
-      }
+      },
     );
   }
 
@@ -141,12 +136,13 @@ class ResultsList extends React.Component {
       result.placeOfPublication[0] : '';
     const items = LibraryItem.getItems(result);
     const totalItems = items.length;
+    const hasRequestTable = items.length === 1;
 
     return (
-      <li key={i} className="nypl-results-item">
+      <li key={i} className={`nypl-results-item ${hasRequestTable ? 'has-request' : ''}`}>
         <h3>
           <Link
-            onClick={(e) => this.getBibRecord(e, bibId, bibTitle)}
+            onClick={e => this.getBibRecord(e, bibId, bibTitle)}
             to={`${appConfig.baseUrl}/bib/${bibId}?searchKeywords=${this.props.searchKeywords}`}
             className="title"
           >
@@ -165,7 +161,7 @@ class ResultsList extends React.Component {
           </ul>
         </div>
         {
-          (items.length === 1) &&
+          hasRequestTable &&
             <ItemTable
               items={items}
               bibId={bibId}
