@@ -16,7 +16,7 @@ class ResultsCount extends React.Component {
   displayContext() {
     const {
       searchKeywords,
-      selectedFacets,
+      selectedFilters,
       field,
     } = this.props;
     const keyMapping = {
@@ -35,15 +35,19 @@ class ResultsCount extends React.Component {
       if (field && field !== 'all') {
         result += `for ${keyMapping[field]} "${searchKeywords}"`;
       } else {
-        result += `for keyword "${searchKeywords}"`;
+        if (searchKeywords.indexOf(' ') !== -1) {
+          result += `for keywords "${searchKeywords}"`;
+        } else {
+          result += `for keyword "${searchKeywords}"`;
+        }
       }
     }
 
-    if (!_isEmpty(selectedFacets)) {
-      _mapObject(selectedFacets, (val, key) => {
+    if (!_isEmpty(selectedFilters)) {
+      _mapObject(selectedFilters, (val, key) => {
         const mappedKey = keyMapping[key];
 
-        if (val[0] && val[0].value) {
+        if (val && val[0] && val[0].value && mappedKey) {
           result += `for ${mappedKey} "${val[0].value}"`;
         }
       });
@@ -52,13 +56,35 @@ class ResultsCount extends React.Component {
     return result;
   }
 
+  /*
+   * checkSelectedFilters()
+   * Returns true if there are any selected format or language filters. TODO: add Date.
+   */
+  checkSelectedFilters() {
+    const selectedFilters = this.props.selectedFilters;
+
+    if ((selectedFilters.materialType && selectedFilters.materialType.length) ||
+      (selectedFilters.language && selectedFilters.language.length)) {
+      return true;
+    }
+
+    // Eventually need to add check for date.
+    return false;
+  }
+
   displayCount() {
-    const { count, isLoading, page } = this.props;
+    const {
+      count,
+      isLoading,
+      page,
+      searchKeywords,
+    } = this.props;
     const countF = count ? count.toLocaleString() : '';
     const displayContext = this.displayContext();
     const start = (page - 1) * 50 + 1;
     const end = (page) * 50 > count ? count : (page * 50);
     const currentResultDisplay = `${start}-${end}`;
+    const plural = (searchKeywords && searchKeywords.indexOf(' ') !== -1) ? 's' : '';
 
     if (isLoading) {
       return (<p>Loading...</p>);
@@ -67,7 +93,20 @@ class ResultsCount extends React.Component {
     if (count !== 0) {
       return (<h2>Displaying {currentResultDisplay} of {countF} results {displayContext}</h2>);
     }
-    return (<h2>No results found. Please try another search.</h2>);
+
+    if (this.checkSelectedFilters()) {
+      return (
+        <h2>No results for the keyword{plural} "{searchKeywords}" with the chosen filters. Try
+          a different search or different filters.
+        </h2>
+      );
+    }
+
+    return (
+      <h2>
+        No results for the keyword{plural} "{searchKeywords}". Try a different search.
+      </h2>
+    );
   }
 
   render() {
@@ -77,7 +116,7 @@ class ResultsCount extends React.Component {
       <div
         id="results-description"
         className="nypl-results-summary"
-        aria-live="assertive"
+        aria-live="polite"
         aria-atomic="true"
         role="presentation"
       >
@@ -91,7 +130,7 @@ ResultsCount.propTypes = {
   count: PropTypes.number,
   page: PropTypes.number,
   isLoading: PropTypes.bool,
-  selectedFacets: PropTypes.object,
+  selectedFilters: PropTypes.object,
   searchKeywords: PropTypes.string,
   field: PropTypes.string,
 };
@@ -100,6 +139,12 @@ ResultsCount.defaultProps = {
   count: 0,
   isLoading: false,
   page: 1,
+  selectedFilters: {
+    materialType: [],
+    language: [],
+    dateAfter: {},
+    dateBefore: {},
+  },
 };
 
 export default ResultsCount;
