@@ -5,29 +5,29 @@ import { IndexRoute, Route, Redirect } from 'react-router';
 /*
  * Components
  */
-import Application from '../components/Application/Application.jsx';
-import Home from '../components/Home/Home.jsx';
-import SearchResultsPage from '../components/SearchResultsPage/SearchResultsPage.jsx';
-import BibPage from '../components/BibPage/BibPage.jsx';
-import HoldRequest from '../components/HoldRequest/HoldRequest.jsx';
-import HoldConfirmation from '../components/HoldConfirmation/HoldConfirmation.jsx';
-import ElectronicDelivery from '../components/ElectronicDelivery/ElectronicDelivery.jsx';
-import NotFound404 from '../components/NotFound404/NotFound404.jsx';
-import appConfig from '../../../appConfig.js';
+import Application from '../components/Application/Application';
+import Home from '../components/Home/Home';
+import SearchResultsPage from '../components/SearchResultsPage/SearchResultsPage';
+import BibPage from '../components/BibPage/BibPage';
+import HoldRequest from '../components/HoldRequest/HoldRequest';
+import HoldConfirmation from '../components/HoldConfirmation/HoldConfirmation';
+import ElectronicDelivery from '../components/ElectronicDelivery/ElectronicDelivery';
+import NotFound404 from '../components/NotFound404/NotFound404';
+import appConfig from '../../../appConfig';
 
 const baseUrl = appConfig.baseUrl;
 const routes = {
   // Routes used in the Express server:
   server: (
-    <Route path={'/'} component={Application}>
+    <Route path="/" component={Application}>
       <IndexRoute component={Home} />
-      <Route path={'/search'} component={SearchResultsPage} />
-      <Route path={'/bib/:bibId'} component={BibPage} />
-      <Route path={'/bib/:bibId/all'} component={BibPage} />
-      <Route path={'/hold/request/:bibId-:itemId'} component={HoldRequest} />
-      <Route path={'/hold/request/:bibId-:itemId/edd'} component={ElectronicDelivery} />
-      <Route path={'/hold/confirmation/:bibId-:itemId'} component={HoldConfirmation} />
-      <Route path={'/404'} component={NotFound404} />
+      <Route path="/search" component={SearchResultsPage} />
+      <Route path="/bib/:bibId" component={BibPage} />
+      <Route path="/bib/:bibId/all" component={BibPage} />
+      <Route path="/hold/request/:bibId-:itemId" component={HoldRequest} />
+      <Route path="/hold/request/:bibId-:itemId/edd" component={ElectronicDelivery} />
+      <Route path="/hold/confirmation/:bibId-:itemId" component={HoldConfirmation} />
+      <Route path="/404" component={NotFound404} />
       <Redirect from="*" to="/404" />
     </Route>
   ),
