@@ -107,21 +107,21 @@ function renderServerValidationError(object) {
 function validate(form, cb) {
   const fieldsToCheck = {
     emailAddress: {
-      validate: (val) => (val.trim().length && isEmail('' + val)),
+      validate: val => (val.trim().length && isEmail('' + val)),
       errorMsg: 'Enter a valid email address. Your request will be delivered to ' +
         'the email address you enter above.',
     },
     chapterTitle: {
-      validate: (val) => !!(val.trim().length),
+      validate: val => !!(val.trim().length),
       errorMsg: 'Indicate the title of the chapter or article you are requesting. ' +
         'Enter "none" if you are requesting an entire item.',
     },
     startPage: {
-      validate: (val) => !!(val.trim().length),
+      validate: val => !!(val.trim().length),
       errorMsg: 'Enter a page number. You may request a maximum of 50 pages.',
     },
     endPage: {
-      validate: (val) => !!(val.trim().length),
+      validate: val => !!(val.trim().length),
       errorMsg: 'Enter a page number. You may request a maximum of 50 pages.',
     },
   };
