@@ -14,9 +14,10 @@ import {
   extend as _extend,
   chain as _chain,
   flatten as _flatten,
+  sortBy as _sortBy,
 } from 'underscore';
 
-import appConfig from '../../../appConfig.js';
+import appConfig from '../../../appConfig';
 
 /**
  * ajaxCall
@@ -28,7 +29,7 @@ import appConfig from '../../../appConfig.js';
 const ajaxCall = (
   endpoint,
   cb = () => {},
-  errorcb = (error) => console.error('Error making ajaxCall', error)
+  errorcb = error => console.error('Error making ajaxCall', error),
 ) => {
   if (!endpoint) return null;
 
@@ -39,18 +40,18 @@ const ajaxCall = (
 };
 
 /**
- * getDefaultFacets
- * Get the default facets needed from the API.
+ * getDefaultFilters
+ * Get the default filters needed from the API.
  * @return {object}
  */
-const getDefaultFacets = () => _extend({}, appConfig.defaultFacets);
+const getDefaultFilters = () => _extend({}, appConfig.defaultFilters);
 
 /**
  * createAppHistory
  * Create a history in the browser or server that coincides with react-router.
  */
 const createAppHistory = () => {
-  if (typeof(window) !== 'undefined') {
+  if (typeof window !== 'undefined') {
     return useQueries(createHistory)();
   }
 
@@ -58,56 +59,61 @@ const createAppHistory = () => {
 };
 
 /**
- * ajaxCall
- * Utility function to make ajax requests.
- * @param {string} endpoint The endpoint to call.
- * @param {function} cb The callback function.
- * @param {function} errorcb The error callback function.
+ * destructureFilters
+ * Get filters directly from the URL and parse and combine them into selected filter values.
+ * @param {object} filters Filters in the url.
+ * @param {object} apiFilters All filters from the API.
  */
-function destructureFilters(filters, apiFacet) {
-  const selectedFacets = {};
-  const facetArray = apiFacet && apiFacet.itemListElement && apiFacet.itemListElement.length ?
-    apiFacet.itemListElement : [];
+function destructureFilters(filters, apiFilters) {
+  const selectedFilters = {};
+  const filterArrayfromAPI =
+    apiFilters && apiFilters.itemListElement && apiFilters.itemListElement.length ?
+      apiFilters.itemListElement : [];
 
   _forEach(filters, (value, key) => {
-    const id = key.substring(8, key.length - 1);
+    const id = key.indexOf('date') !== -1 ?
+      // Because filters are in the form of `filters[language][0]`;
+      key.substring(8, key.length - 1) : key.substring(8, key.length - 4);
 
     if (id === 'dateAfter' || id === 'dateBefore') {
-      selectedFacets[id] = {
-        id: value,
-        value: id === 'dateAfter' ? `after ${value}` : `before ${value}`,
+      selectedFilters[id] = {
+        label: value,
+        value,
       };
     } else if (_isArray(value) && value.length) {
-      if (!selectedFacets[id]) {
-        selectedFacets[id] = [];
+      if (!selectedFilters[id]) {
+        selectedFilters[id] = [];
       }
-      _forEach(value, facetValue => {
-        const facetObjFromAPI = _findWhere(facetArray, { id });
-        if (facetObjFromAPI && facetObjFromAPI.values && facetObjFromAPI.values.length) {
-          const facet = _findWhere(facetObjFromAPI.values, { value: facetValue });
-          if (facet) {
-            selectedFacets[id].push({
-              id: facet.value,
-              value: facet.label || facet.value,
+      _forEach(value, (filterValue) => {
+        const filterObjFromApi = _findWhere(filterArrayfromAPI, { id });
+        if (filterObjFromApi && filterObjFromApi.values && filterObjFromApi.values.length) {
+          const filter = _findWhere(filterObjFromApi.values, { value: filterValue });
+          if (filter) {
+            selectedFilters[id].push({
+              value: filter.value,
+              label: filter.label || filter.value,
             });
           }
         }
       });
     } else if (typeof value === 'string') {
-      const facetObjFromAPI = _findWhere(facetArray, { id });
-      if (facetObjFromAPI && facetObjFromAPI.values && facetObjFromAPI.values.length) {
-        const facet = _findWhere(facetObjFromAPI.values, { value });
-        if (facet) {
-          selectedFacets[id] = [{
-            id: facet.value,
-            value: facet.label || facet.value,
-          }];
+      const filterObjFromApi = _findWhere(filterArrayfromAPI, { id });
+      if (filterObjFromApi && filterObjFromApi.values && filterObjFromApi.values.length) {
+        const filter = _findWhere(filterObjFromApi.values, { value });
+        if (filter) {
+          if (!selectedFilters[id]) {
+            selectedFilters[id] = [];
+          }
+          selectedFilters[id].push({
+            value: filter.value,
+            label: filter.label || filter.value,
+          });
         }
       }
     }
   });
 
-  return selectedFacets;
+  return selectedFilters;
 }
 
 /**
@@ -128,31 +134,27 @@ const getSortQuery = (sortBy = '') => {
 };
 
 /**
- * getFacetFilterParam
- * Get the search params from the facet values.
- * @param {object} facets Key/value pair of facet and the selected value.
+ * getFilterParam
+ * Get the search params from the filter values.
+ * @param {object} filters Key/value pair of filter and the selected value.
  */
-const getFacetFilterParam = (facets) => {
+const getFilterParam = (filters) => {
   let strSearch = '';
 
-  if (!_isEmpty(facets)) {
-    _mapObject(facets, (val, key) => {
-      // Property contains an array of its selected facet values:
-      if (val.length && _isArray(val)) {
-        _forEach(val, (facet) => {
-          if (facet.value && facet.value !== '') {
-            // At this time, materialType filter requires _packed for filtering (but not as data).
-            // Other filters do not.
-            strSearch += (key === 'materialType' ?
-              `&filters[${key}_packed]=${encodeURIComponent(facet.id)}` :
-              `&filters[${key}]=${encodeURIComponent(facet.id)}`);
-          } else if (typeof facet === 'string') {
-            strSearch += `&filters[${key}]=${encodeURIComponent(facet)}`;
+  if (!_isEmpty(filters)) {
+    _mapObject(filters, (val, key) => {
+      // Property contains an array of its selected filter values:
+      if (val && val.length && _isArray(val)) {
+        _forEach(val, (filter, index) => {
+          if (filter.value && filter.value !== '') {
+            strSearch += `&filters[${key}][${index}]=${encodeURIComponent(filter.value)}`;
+          } else if (typeof filter === 'string') {
+            strSearch += `&filters[${key}][${index}]=${encodeURIComponent(filter)}`;
           }
         });
-      } else if (val.value && val.value !== '') {
-        strSearch += `&filters[${key}]=${encodeURIComponent(val.id)}`;
-      } else if (typeof val === 'string') {
+      } else if (val && val.value && val.value !== '') {
+        strSearch += `&filters[${key}]=${encodeURIComponent(val.value)}`;
+      } else if (val && typeof val === 'string') {
         strSearch += `&filters[${key}]=${encodeURIComponent(val)}`;
       }
     });
@@ -189,29 +191,30 @@ const trackDiscovery = gaUtils.trackEvent('Discovery');
  * @example
  * const apiQueryFunc = basicQuery(this.props);
  * const apiQuery = apiQueryFunc();
- * // apiQuery == 'q='
+ * // apiQuery === 'q='
  * const apiQuery2 = apiQueryFunc({ page: 3 });
- * // apiQuery2 == 'q=&page=3'
+ * // apiQuery2 === 'q=&page=3'
  * const apiQuery3 = apiQueryFunc({ page: 3, q: 'hamlet' });
- * // apiQuery3 == 'q=hamlet&page=3'
+ * // apiQuery3 === 'q=hamlet&page=3'
  * @param {object} props The application props.
  */
 const basicQuery = (props = {}) => {
   return ({
     sortBy,
     field,
-    selectedFacets,
+    selectedFilters,
     searchKeywords,
     page,
   }) => {
     const sortQuery = getSortQuery(sortBy || props.sortBy);
     const fieldQuery = getFieldParam(field || props.field);
-    const filterQuery = getFacetFilterParam(selectedFacets || props.selectedFacets);
+    const filterQuery = getFilterParam(selectedFilters || props.selectedFilters);
     // `searchKeywords` can be an empty string, so check if it's undefined instead.
     const query = searchKeywords !== undefined ? searchKeywords : props.searchKeywords;
     const searchKeywordsQuery = query ? encodeURIComponent(query) : '';
     let pageQuery = props.page && props.page !== '1' ? `&page=${props.page}` : '';
     pageQuery = page && page !== '1' ? `&page=${page}` : pageQuery;
+    pageQuery = page === '1' ? '' : pageQuery;
 
     return `q=${searchKeywordsQuery}${filterQuery}${sortQuery}${fieldQuery}${pageQuery}`;
   };
@@ -221,7 +224,7 @@ const basicQuery = (props = {}) => {
  * getReqParams
  * Read the query param from the request object from Express and returns its value or
  * default values for each. It also returns a string representation of all the selected
- * facets in the url from the `filter` query.
+ * filters in the url from the `filter` query.
  * @param {object} query The request query object from Express.
  */
 function getReqParams(query = {}) {
@@ -261,31 +264,42 @@ function getReqParams(query = {}) {
  * @return {object}
  */
 function parseServerSelectedFilters(filters, dateAfter, dateBefore) {
-  const selectedFacets = {};
+  const selectedFilters = {
+    materialType: [],
+    language: [],
+    dateAfter: {},
+    dateBefore: {},
+  };
+
   if (_isArray(filters) && filters.length && !_isEmpty(filters[0])) {
     _chain(filters)
       // Each incoming filter is in JSON string format so it needs to be parsed first.
       .map(filter => JSON.parse(filter))
-      // Group selected facets into arrays according to their field.
+      // Group selected filters into arrays according to their field.
       .groupBy('field')
       // Created the needed data structure.
-      .mapObject((facetArray, key) => {
+      .mapObject((filterArray, key) => {
         if (key) {
-          selectedFacets[key] =
-            facetArray.map((facet) => ({ id: facet.value, value: facet.label }));
+          selectedFilters[key] =
+            filterArray.map(filter => ({
+              value: filter.value,
+              label: filter.label,
+              count: filter.count,
+              selected: true,
+            }));
         }
       });
   }
 
   if (dateAfter) {
-    selectedFacets.dateAfter = { id: dateAfter, value: dateAfter };
+    selectedFilters.dateAfter = { id: dateAfter, value: dateAfter };
   }
 
   if (dateBefore) {
-    selectedFacets.dateBefore = { id: dateBefore, value: dateBefore };
+    selectedFilters.dateBefore = { id: dateBefore, value: dateBefore };
   }
 
-  return selectedFacets;
+  return selectedFilters;
 }
 
 /**
@@ -301,7 +315,7 @@ function getAggregatedElectronicResources(items = []) {
 
   const electronicResources = [];
 
-  _forEach(items, item => {
+  _forEach(items, (item) => {
     if (item.isElectronicResource) {
       electronicResources.push(item.electronicResources);
     }
@@ -310,17 +324,55 @@ function getAggregatedElectronicResources(items = []) {
   return _flatten(electronicResources);
 }
 
+/**
+ * getUpdatedFilterValues(props)
+ * Get an array of filter values based on one filter. If any filters are selected, they'll
+ * be `true` in their object property `selected`.
+ * @param {object} props
+ * @return {array}
+ */
+const getUpdatedFilterValues = (props) => {
+  const {
+    filter,
+    selectedFilters,
+  } = props;
+  const filterValues = filter && filter.values && filter.values.length ? filter.values : [];
+  // Just want to add the `selected` property here.
+  const defaultFilterValues = filterValues.map(value => _extend({ selected: false }, value));
+  let updatedFilterValues = defaultFilterValues;
+
+  // If there are selected filters, then we want to update the filter values with those
+  // filters already selected. That way, the checkboxes will be checked.
+  if (selectedFilters) {
+    updatedFilterValues = defaultFilterValues.map((defaultFilterValue) => {
+      const defaultFilter = defaultFilterValue;
+      selectedFilters.forEach((selectedFilter) => {
+        if (selectedFilter.value === defaultFilter.value) {
+          defaultFilter.selected = true;
+        }
+      });
+
+      return defaultFilter;
+    });
+  }
+
+  updatedFilterValues = _sortBy(updatedFilterValues, f => f.label);
+
+  return updatedFilterValues;
+};
+
 export {
   trackDiscovery,
   ajaxCall,
   getSortQuery,
   createAppHistory,
   getFieldParam,
-  getFacetFilterParam,
+  getFilterParam,
   destructureFilters,
-  getDefaultFacets,
+  getDefaultFilters,
   basicQuery,
   getReqParams,
   parseServerSelectedFilters,
   getAggregatedElectronicResources,
+  getUpdatedFilterValues,
 };
