@@ -1,18 +1,18 @@
 import express from 'express';
 
-import Bib from './Bib.js';
-import Hold from './Hold.js';
-import Search from './Search.js';
-import appConfig from '../../../appConfig.js';
+import Bib from './Bib';
+import Hold from './Hold';
+import Search from './Search';
+import appConfig from '../../../appConfig';
 
 const router = express.Router();
 
 function MainApp(req, res, next) {
   res.locals.data.Store = {
     searchResults: {},
-    selectedFacets: {},
+    selectedFilters: {},
     searchKeywords: '',
-    facets: {},
+    filters: {},
     page: '1',
     sortBy: 'relevance',
     field: 'all',
