@@ -1,14 +1,14 @@
 import nyplApiClient from '../routes/nyplApiClient';
 import logger from '../../../logger';
-import appConfig from '../../../appConfig.js';
+import appConfig from '../../../appConfig';
 
-const nyplApiClientCall = (query) =>
+const nyplApiClientCall = query =>
   nyplApiClient().then(client => client.get(`/discovery/resources/${query}`, { cache: false }));
 
 function fetchBib(bibId, cb, errorcb) {
   return nyplApiClientCall(bibId)
     .then(response => cb(response))
-    .catch(error => {
+    .catch((error) => {
       logger.error(`Error attemping to fetch a Bib server side in fetchBib, id: ${bibId}`, error);
 
       errorcb(error);
@@ -40,7 +40,7 @@ function bibSearchServer(req, res, next) {
         error,
       };
       next();
-    }
+    },
   );
 }
 
@@ -49,8 +49,8 @@ function bibSearchAjax(req, res) {
 
   fetchBib(
     bibId,
-    (data) => res.json(data),
-    (error) => res.json(error)
+    data => res.json(data),
+    error => res.json(error),
   );
 }
 
