@@ -3,9 +3,11 @@ import React from 'react';
 import sinon from 'sinon';
 import { expect } from 'chai';
 import { mount } from 'enzyme';
+
 // Import the component that is going to be tested
-import HoldConfirmation from './../../src/app/components/HoldConfirmation/HoldConfirmation.jsx';
-import Actions from './../../src/app/actions/Actions.js';
+import HoldConfirmation from './../../src/app/components/HoldConfirmation/HoldConfirmation';
+import Actions from './../../src/app/actions/Actions';
+import appConfig from '../../appConfig';
 
 describe('HoldConfirmation', () => {
   describe('After being rendered, <HoldConfirmation>', () => {
@@ -20,7 +22,7 @@ describe('HoldConfirmation', () => {
 
     before(() => {
       requireUser = sinon.spy(HoldConfirmation.prototype, 'requireUser');
-      component = mount(<HoldConfirmation location={location} />);
+      component = mount(<HoldConfirmation location={location} />, { attachTo: document.body });
     });
 
     after(() => {
@@ -46,7 +48,7 @@ describe('HoldConfirmation', () => {
     before(() => {
       Actions.updatePatronData({});
       requireUser = sinon.spy(HoldConfirmation.prototype, 'requireUser');
-      component = mount(<HoldConfirmation location={location} />);
+      component = mount(<HoldConfirmation location={location} />, { attachTo: document.body });
     });
 
     after(() => {
@@ -62,8 +64,6 @@ describe('HoldConfirmation', () => {
 
   describe('If the patron is logged in but the App doesn\'t get valid data,<HoldConfirmation>',
     () => {
-      let component;
-      let requireUser;
       const location = {
         query: {
           pickupLocation: 'myr',
@@ -72,6 +72,8 @@ describe('HoldConfirmation', () => {
           errorMessage: 'Something is wrong with the server.',
         },
       };
+      let component;
+      let requireUser;
 
       before(() => {
         Actions.updatePatronData({
@@ -80,7 +82,7 @@ describe('HoldConfirmation', () => {
           barcodes: ['162402680435300'],
         });
         requireUser = sinon.spy(HoldConfirmation.prototype, 'requireUser');
-        component = mount(<HoldConfirmation location={location} />);
+        component = mount(<HoldConfirmation location={location} />, { attachTo: document.body });
       });
 
       after(() => {
@@ -99,26 +101,21 @@ describe('HoldConfirmation', () => {
           <p>
             We could not process your request at this time. Please try again or contact 917-ASK-NYPL
             (<a href="tel:19172756975">917-275-6975</a>).
-          </p>
+          </p>,
         )).to.equal(true);
       });
-    }
+    },
   );
 
   describe('If the patron is logged in and the App receives valid data, <HoldConfirmation>', () => {
-    let component;
-    let requireUser;
-    let modelDeliveryLocationName;
     const location = {
       query: {
         pickupLocation: 'mala',
       },
     };
-
     const bib = {
       title: ['Harry Potter'],
     };
-
     const deliveryLocations = [
       {
         '@id': 'loc:myr',
@@ -139,6 +136,10 @@ describe('HoldConfirmation', () => {
         shortName: 'Schwarzman Building',
       },
     ];
+    let component;
+    let requireUser;
+    let modelDeliveryLocationName;
+    let pushSpy;
 
     before(() => {
       Actions.updatePatronData({
@@ -149,14 +150,19 @@ describe('HoldConfirmation', () => {
       requireUser = sinon.spy(HoldConfirmation.prototype, 'requireUser');
       modelDeliveryLocationName = sinon.spy(
         HoldConfirmation.prototype,
-        'modelDeliveryLocationName'
+        'modelDeliveryLocationName',
       );
+      pushSpy = sinon.spy();
       component = mount(
         <HoldConfirmation
           location={location}
           bib={bib}
           deliveryLocations={deliveryLocations}
-        />
+        />,
+        {
+          context: { router: { createHref: () => {}, push: pushSpy } },
+          attachTo: document.body,
+        },
       );
     });
 
@@ -174,7 +180,9 @@ describe('HoldConfirmation', () => {
       expect(pageHeader).to.have.length(1);
       expect(pageHeader.find('h1')).to.have.length(1);
       expect(pageHeader.find('h1').text()).to.equal('Request Confirmation');
-      expect(pageHeader.contains(<h1>Request Confirmation</h1>)).to.equal(true);
+      expect(
+        pageHeader.contains(<h1 id="confirmation-title" tabIndex="0">Request Confirmation</h1>)
+      ).to.equal(true);
     });
 
     it('should display the layout of request confirmation page\'s contents.', () => {
@@ -191,14 +199,12 @@ describe('HoldConfirmation', () => {
     it('should render the message for the physical delivery location.', () => {
       const main = component.find('main');
 
-      expect(modelDeliveryLocationName.returnValues[0]).to.equal(
-        'Schwarzman Building - Allen Scholar Room'
-      );
+      expect(modelDeliveryLocationName.returnValues[0])
+        .to.equal('Schwarzman Building - Allen Scholar Room');
 
       expect(main.find('#delivery-location')).to.have.length(1);
-      expect(main.find('#delivery-location').text()).to.equal(
-        'The item will be delivered to: Schwarzman Building - Allen Scholar Room'
-      );
+      expect(main.find('#delivery-location').text())
+        .to.equal('The item will be delivered to: Schwarzman Building - Allen Scholar Room');
     });
 
     it('should deliver the item\'s title on the page.', () => {
@@ -213,6 +219,58 @@ describe('HoldConfirmation', () => {
 
       expect(main.find('#start-new-search')).to.have.length(1);
       expect(main.find('#start-new-search').text()).to.equal('Start a new search');
+    });
+
+    it('should call the React Router context when the link to the homepage is clicked.', () => {
+      const main = component.find('main');
+
+      main.find('#start-new-search').simulate('click');
+      expect(pushSpy.callCount).to.equal(1);
+      expect(pushSpy.calledWith(`${appConfig.baseUrl}/`)).to.equal(true);
+    });
+  });
+
+  describe('If the patron is logged in and the App receives valid data but no delivery ' +
+    'location label', () => {
+    const location = {
+      query: {
+        pickupLocation: 'myr',
+      },
+    };
+    const deliveryLocations = [
+      {
+        '@id': 'loc:myr',
+        address: '',
+        prefLabel: '',
+        shortName: '',
+      },
+    ];
+    let component;
+    let modelDeliveryLocationName;
+
+    before(() => {
+      Actions.updatePatronData({
+        id: '6677200',
+        names: ['Leonard, Mike'],
+        barcodes: ['162402680435300'],
+      });
+      modelDeliveryLocationName = sinon.spy(
+        HoldConfirmation.prototype,
+        'modelDeliveryLocationName',
+      );
+      component = mount(
+        <HoldConfirmation location={location} deliveryLocations={deliveryLocations} />,
+        { attachTo: document.body }
+      );
+    });
+
+    after(() => {
+      modelDeliveryLocationName.restore();
+      component.unmount();
+    });
+
+    it('should display the layout of request confirmation page\'s header.', () => {
+      expect(modelDeliveryLocationName.returnValues[0]).to.equal('');
     });
   });
 
@@ -262,7 +320,8 @@ describe('HoldConfirmation', () => {
           location={location}
           bib={bib}
           deliveryLocations={deliveryLocations}
-        />);
+        />,
+        { attachTo: document.body });
     });
 
     after(() => {
@@ -306,7 +365,8 @@ describe('HoldConfirmation', () => {
         barcodes: ['162402680435300'],
       });
       requireUser = sinon.spy(HoldConfirmation.prototype, 'requireUser');
-      component = mount(<HoldConfirmation location={location} bib={bib} />);
+      component =
+        mount(<HoldConfirmation location={location} bib={bib} />, { attachTo: document.body });
     });
 
     after(() => {
@@ -314,7 +374,7 @@ describe('HoldConfirmation', () => {
       component.unmount();
     });
 
-    it('should render the defalut error message.', () => {
+    it('should render the default error message.', () => {
       const main = component.find('main');
 
       expect(main.find('#delivery-location')).to.have.length(1);
@@ -376,7 +436,8 @@ describe('HoldConfirmation', () => {
           location={location}
           bib={bib}
           deliveryLocations={deliveryLocations}
-        />
+        />,
+        { attachTo: document.body }
       );
     });
 
@@ -389,9 +450,8 @@ describe('HoldConfirmation', () => {
       const main = component.find('main');
 
       expect(main.find('#delivery-location')).to.have.length(1);
-      expect(main.find('#delivery-location').text()).to.equal(
-        'The item will be delivered to: n/a (electronic delivery)'
-      );
+      expect(main.find('#delivery-location').text())
+        .to.equal('The item will be delivered to: n/a (electronic delivery)');
     });
   });
 
@@ -408,7 +468,8 @@ describe('HoldConfirmation', () => {
     };
 
     before(() => {
-      component = mount(<HoldConfirmation location={location} bib={bib} />);
+      component =
+        mount(<HoldConfirmation location={location} bib={bib} />, { attachTo: document.body });
     });
 
     after(() => {
@@ -425,20 +486,27 @@ describe('HoldConfirmation', () => {
   });
 
   describe('If the patron gets here from a search result page, <HoldConfirmation>', () => {
-    let component;
     const location = {
       query: {
         pickupLocation: 'myr',
         searchKeywords: 'Bryant',
       },
     };
-
     const bib = {
       title: ['Harry Potter'],
     };
+    let component;
+    let pushSpy;
 
     before(() => {
-      component = mount(<HoldConfirmation location={location} bib={bib} />);
+      pushSpy = sinon.spy();
+      component = mount(
+        <HoldConfirmation location={location} bib={bib} searchKeywords="Bryant" />,
+        {
+          context: { router: { createHref: () => {}, push: pushSpy } },
+          attachTo: document.body,
+        },
+      );
     });
 
     after(() => {
@@ -451,56 +519,57 @@ describe('HoldConfirmation', () => {
       expect(main.find('#start-new-search')).to.have.length(1);
       expect(main.find('#start-new-search').text()).to.equal('start a new search');
       expect(main.find('#go-back-search-results')).to.have.length(1);
-      expect(main.find('#go-back-search-results').text()).to.equal(
-        'Go back to your search results'
-      );
+      expect(main.find('#go-back-search-results').text())
+        .to.equal('Go back to your search results');
+    });
+
+    it('should call the React Router context with the link back to the search results', () => {
+      const main = component.find('main');
+
+      main.find('#go-back-search-results').simulate('click');
+      expect(pushSpy.callCount).to.equal(1);
+      expect(pushSpy.calledWith(`${appConfig.baseUrl}/search?q=Bryant`)).to.equal(true);
     });
   });
 
-  describe('If the patron does not get here from a classic catalog, <HoldConfirmation>',
-    () => {
-      let component;
-      let renderBackToClassicLink;
-      const location = {
-        query: {
-          pickupLocation: 'myr',
-          searchKeywords: 'Bryant',
-        },
-      };
+  describe('If the patron does not get here from a classic catalog, <HoldConfirmation>', () => {
+    let component;
+    let renderBackToClassicLink;
+    const location = {
+      query: {
+        pickupLocation: 'myr',
+        searchKeywords: 'Bryant',
+      },
+    };
 
-      const bib = {
-        title: ['Harry Potter'],
-      };
+    const bib = {
+      title: ['Harry Potter'],
+    };
 
-      before(() => {
-        renderBackToClassicLink = sinon.spy(HoldConfirmation.prototype, 'renderBackToClassicLink');
-        component = mount(<HoldConfirmation location={location} bib={bib} />);
-      });
+    before(() => {
+      renderBackToClassicLink = sinon.spy(HoldConfirmation.prototype, 'renderBackToClassicLink');
+      component =
+        mount(<HoldConfirmation location={location} bib={bib} />, { attachTo: document.body });
+    });
 
-      after(() => {
-        renderBackToClassicLink.restore();
-        component.unmount();
-      });
+    after(() => {
+      renderBackToClassicLink.restore();
+      component.unmount();
+    });
 
-      it('should not have the link back to the classic catalog homepage.',
-        () => {
-          const main = component.find('main');
+    it('should not have the link back to the classic catalog homepage.', () => {
+      const main = component.find('main');
 
-          expect(renderBackToClassicLink.returnValues[0]).to.equal(false);
+      expect(renderBackToClassicLink.returnValues[0]).to.equal(false);
+      expect(main.find('#go-back-catalog')).to.have.length(0);
+    });
 
-          expect(main.find('#go-back-catalog')).to.have.length(0);
-        }
-      );
+    it('should have the link to shared collection catalog.', () => {
+      const main = component.find('main');
 
-      it('should have the link to shared collection catalog.', () => {
-        const main = component.find('main');
-
-        expect(main.find('#start-new-search').text()).to.equal(
-          'start a new search'
-        );
-      });
-    }
-  );
+      expect(main.find('#start-new-search').text()).to.equal('start a new search');
+    });
+  });
 
   describe('If the patron gets here from a classic catalog search result page, <HoldConfirmation>',
     () => {
@@ -520,7 +589,8 @@ describe('HoldConfirmation', () => {
       };
 
       before(() => {
-        component = mount(<HoldConfirmation location={location} bib={bib} />);
+        component =
+          mount(<HoldConfirmation location={location} bib={bib} />, { attachTo: document.body });
       });
 
       after(() => {
@@ -544,16 +614,15 @@ describe('HoldConfirmation', () => {
           expect(main.find('#go-back-catalog a').at(1).text()).to.equal('start a new search');
           expect(main.find('#go-back-catalog a').at(1).prop('href'))
             .to.equal('https://catalog.nypl.org/search');
-        }
+        },
       );
 
       it('should have the link to shared collection catalog.', () => {
         const main = component.find('main');
 
-        expect(main.find('#go-to-shared-catalog').text()).to.equal(
-          ' You may also try your search in our Shared Collection Catalog.'
-        );
+        expect(main.find('#go-to-shared-catalog').text())
+          .to.equal(' You may also try your search in our Shared Collection Catalog.');
       });
-    }
+    },
   );
 });
