@@ -4,8 +4,8 @@ import sinon from 'sinon';
 import { expect } from 'chai';
 import { mount } from 'enzyme';
 // Import the component that is going to be tested
-import HoldRequest from './../../src/app/components/HoldRequest/HoldRequest.jsx';
-import Actions from './../../src/app/actions/Actions.js';
+import HoldRequest from './../../src/app/components/HoldRequest/HoldRequest';
+import Actions from './../../src/app/actions/Actions';
 
 const mockedItems = [
   {
@@ -72,7 +72,7 @@ describe('HoldRequest', () => {
 
     before(() => {
       requireUser = sinon.spy(HoldRequest.prototype, 'requireUser');
-      component = mount(<HoldRequest />);
+      component = mount(<HoldRequest />, { attachTo: document.body });
     });
 
     after(() => {
@@ -93,7 +93,7 @@ describe('HoldRequest', () => {
     before(() => {
       Actions.updatePatronData({});
       requireUser = sinon.spy(HoldRequest.prototype, 'requireUser');
-      component = mount(<HoldRequest />);
+      component = mount(<HoldRequest />, { attachTo: document.body });
     });
 
     after(() => {
@@ -112,7 +112,7 @@ describe('HoldRequest', () => {
       let component;
 
       before(() => {
-        component = mount(<HoldRequest />);
+        component = mount(<HoldRequest />, { attachTo: document.body });
       });
 
       after(() => {
@@ -137,7 +137,7 @@ describe('HoldRequest', () => {
           </h2>)
         ).to.equal(true);
       });
-    }
+    },
   );
 
   describe('If the patron is logged in and the App receives invalid delivery location data, ' +
@@ -150,7 +150,7 @@ describe('HoldRequest', () => {
     };
 
     before(() => {
-      component = mount(<HoldRequest bib={bib} params={{ itemId: 'i10000003' }} />);
+      component = mount(<HoldRequest bib={bib} params={{ itemId: 'i10000003' }} />, { attachTo: document.body });
     });
 
     after(() => {
@@ -169,7 +169,7 @@ describe('HoldRequest', () => {
 
       expect(form.find('h2')).to.have.length(1);
       expect(form.contains(
-        <h2>
+        <h2 className="nypl-request-form-title">
           Delivery options for this item are currently unavailable. Please try again later or
           contact 917-ASK-NYPL (<a href="tel:917-275-6975">917-275-6975</a>).
         </h2>
@@ -213,7 +213,8 @@ describe('HoldRequest', () => {
           bib={bib}
           deliveryLocations={deliveryLocations}
           params={{ itemId: 'i10000003' }}
-        />
+        />,
+        { attachTo: document.body }
       );
     });
 
@@ -226,7 +227,7 @@ describe('HoldRequest', () => {
 
       expect(form.find('h2')).to.have.length(1);
       expect(form.contains(
-        <h2>Choose a delivery option or location</h2>
+        <h2 className="nypl-request-form-title">Choose a delivery option or location</h2>
       )).to.equal(true);
     });
 
@@ -259,9 +260,7 @@ describe('HoldRequest', () => {
       const label2 = fieldset.find('label').at(2);
 
       expect(label0.find('.nypl-screenreader-only').text()).to.equal('Send to:');
-      expect(label0.find('.nypl-location-name').text()).to.equal(
-        'Library for the Performing Arts'
-      );
+      expect(label0.find('.nypl-location-name').text()).to.equal('Library for the Performing Arts');
       expect(label0.find('.nypl-location-address').text()).to.equal('40 Lincoln Center Plaza');
 
       expect(label1.find('.nypl-screenreader-only').text()).to.equal('Send to:');
@@ -269,12 +268,10 @@ describe('HoldRequest', () => {
       expect(label1.find('.nypl-location-address').text()).to.equal('515 Malcolm X Boulevard');
 
       expect(label2.find('.nypl-screenreader-only').text()).to.equal('Send to:');
-      expect(label2.find('.nypl-location-name').text()).to.equal(
-        'Schwarzman Building - Allen Scholar Room'
-      );
-      expect(label2.find('.nypl-location-address').text()).to.equal(
-        '476 Fifth Avenue (42nd St and Fifth Ave)'
-      );
+      expect(label2.find('.nypl-location-name').text())
+        .to.equal('Schwarzman Building - Allen Scholar Room');
+      expect(label2.find('.nypl-location-address').text())
+        .to.equal('476 Fifth Avenue (42nd St and Fifth Ave)');
     });
 
     it('should deliver request button with the respective URL on the page.', () => {
@@ -322,7 +319,8 @@ describe('HoldRequest', () => {
           deliveryLocations={deliveryLocations}
           isEddRequestable
           params={{ itemId: 'i10000003' }}
-        />
+        />,
+        { attachTo: document.body }
       );
     });
 
@@ -339,9 +337,8 @@ describe('HoldRequest', () => {
       expect(fieldset.find('legend')).to.have.length(1);
       expect(fieldset.find('label').at(0).find('input').props().type).to.equal('radio');
       expect(fieldset.find('label').at(0).find('input').props().checked).to.equal(true);
-      expect(fieldset.find('label').at(0).text()).to.equal(
-        'Have up to 50 pages scanned and sent to you via electronic mail.'
-      );
+      expect(fieldset.find('label').at(0).text())
+        .to.equal('Have up to 50 pages scanned and sent to you via electronic mail.');
     });
   });
 });
