@@ -4,8 +4,8 @@ const merge = require('webpack-merge');
 const CleanBuild = require('clean-webpack-plugin');
 const ExtractTextPlugin = require('extract-text-webpack-plugin');
 const sassPaths = require('@nypl/design-toolkit').includePaths
-  .map((sassPath) => sassPath).join('&');
-const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
+  .map(sassPath => sassPath).join('&');
+// const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
 
 // References the applications root path
 const ROOT_PATH = path.resolve(__dirname);
@@ -43,39 +43,39 @@ const commonSettings = {
     new webpack.DefinePlugin({
       loadA11y: process.env.loadA11y || false,
     }),
-    new BundleAnalyzerPlugin({
-      // Can be `server`, `static` or `disabled`.
-      // In `server` mode analyzer will start HTTP server to show bundle report.
-      // In `static` mode single HTML file with bundle report will be generated.
-      // In `disabled` mode you can use this plugin to just generate Webpack Stats JSON file by
-      // setting `generateStatsFile` to `true`.
-      analyzerMode: 'static',
-      // Host that will be used in `server` mode to start HTTP server.
-      // analyzerHost: '127.0.0.1',
-      // Port that will be used in `server` mode to start HTTP server.
-      // analyzerPort: 8888,
-      // Path to bundle report file that will be generated in `static` mode.
-      // Relative to bundles output directory.
-      reportFilename: 'report.html',
-      // Module sizes to show in report by default.
-      // Should be one of `stat`, `parsed` or `gzip`.
-      // See "Definitions" section for more information.
-      defaultSizes: 'parsed',
-      // Automatically open report in default browser
-      openAnalyzer: false,
-      // If `true`, Webpack Stats JSON file will be generated in bundles output directory
-      generateStatsFile: false,
-      // Name of Webpack Stats JSON file that will be generated if `generateStatsFile` is `true`.
-      // Relative to bundles output directory.
-      statsFilename: 'stats.json',
-      // Options for `stats.toJson()` method.
-      // For example you can exclude sources of your modules from stats file with `source: false`
-      // option.
-      // See more options here: https://github.com/webpack/webpack/blob/webpack-1/lib/Stats.js#L21
-      statsOptions: null,
-      // Log level. Can be 'info', 'warn', 'error' or 'silent'.
-      logLevel: 'info',
-    }),
+    // new BundleAnalyzerPlugin({
+    //   // Can be `server`, `static` or `disabled`.
+    //   // In `server` mode analyzer will start HTTP server to show bundle report.
+    //   // In `static` mode single HTML file with bundle report will be generated.
+    //   // In `disabled` mode you can use this plugin to just generate Webpack Stats JSON file by
+    //   // setting `generateStatsFile` to `true`.
+    //   analyzerMode: 'static',
+    //   // Host that will be used in `server` mode to start HTTP server.
+    //   // analyzerHost: '127.0.0.1',
+    //   // Port that will be used in `server` mode to start HTTP server.
+    //   // analyzerPort: 8888,
+    //   // Path to bundle report file that will be generated in `static` mode.
+    //   // Relative to bundles output directory.
+    //   reportFilename: 'report.html',
+    //   // Module sizes to show in report by default.
+    //   // Should be one of `stat`, `parsed` or `gzip`.
+    //   // See "Definitions" section for more information.
+    //   defaultSizes: 'parsed',
+    //   // Automatically open report in default browser
+    //   openAnalyzer: false,
+    //   // If `true`, Webpack Stats JSON file will be generated in bundles output directory
+    //   generateStatsFile: false,
+    //   // Name of Webpack Stats JSON file that will be generated if `generateStatsFile` is `true`.
+    //   // Relative to bundles output directory.
+    //   statsFilename: 'stats.json',
+    //   // Options for `stats.toJson()` method.
+    //   // For example you can exclude sources of your modules from stats file with `source: false`
+    //   // option.
+    //   // See more options here: https://github.com/webpack/webpack/blob/webpack-1/lib/Stats.js#L21
+    //   statsOptions: null,
+    //   // Log level. Can be 'info', 'warn', 'error' or 'silent'.
+    //   logLevel: 'info',
+    // }),
   ],
 };
 
@@ -86,7 +86,7 @@ const commonSettings = {
  * the common app configuration with the
  * additional development specific settings.
  *
-**/
+ */
 // Need to configure webpack-dev-server and hot-reload
 // module correctly.
 if (ENV === 'development') {
@@ -138,7 +138,7 @@ if (ENV === 'development') {
  * the common app configuration with the
  * additional production specific settings.
  *
-**/
+ */
 if (ENV === 'production') {
   const loaders = [
     {
